Finishing up

+++ migrations/3_3_up.sql
@@ -2,6 +2,6 @@ CREATE TABLE board.thread_member
 (
     UserId UUID  NOT NULL REFERENCES board.user (Id),
     ThreadId UUID  NOT NULL REFERENCES board.thread (Id),
-    LastViewedPostUnixTime int NOT NULL DEFAULT 0,
+    LastViewedPostUnixTime bigint NOT NULL DEFAULT 0,
     PRIMARY KEY (UserId, ThreadId)
 );
